Cleanup discriminator convention detection
... and a few other matters

Primarily:
- Move discriminator detection and rendering to own .tsx module
- Cleanup naming and types to be more precise

Also:
- Start using json-schema-typed in web
- Fix a broken type import (why were we not alerted?)
- Enable TypeScript's strict: true in web
- Fix strictness and other mistake in CreateNodesWrapper
  - Define type for props to reflect use of `[internalIdKey]` symbol
  - Pass original `props.schema` to @theme-original's <CreateNodes />,
    rather than the mistaken partial ...rest
